# src/main/resources/META-INF/resources/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "AddMemberRequest" : {
        "type" : "object",
        "properties" : {
          "role" : {
            "$ref" : "#/components/schemas/ProjectRole"
          }
        }
      },
      "Bucket" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        }
      },
      "CreateProjectRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "CreateSampleRequest" : {
        "type" : "object",
        "required" : [ "name", "alias", "taxId", "mlst", "isolationSource", "collectionDate", "location", "sequencingLab", "institution", "hostHealthState" ],
        "properties" : {
          "name" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "alias" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "taxId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hostTaxId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "mlst" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "isolationSource" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "collectionDate" : {
            "type" : "string",
            "pattern" : "\\S",
            "$ref" : "#/components/schemas/LocalDate"
          },
          "location" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "sequencingLab" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "institution" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "hostHealthState" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        }
      },
      "File" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "uuid" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "fileName" : {
            "type" : "string"
          },
          "virtualPath" : {
            "type" : "string"
          },
          "objectKey" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/UploadStatus"
          },
          "sample" : {
            "$ref" : "#/components/schemas/Sample"
          }
        }
      },
      "HealthStatusResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        }
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
      "MinioEvent" : {
        "type" : "object",
        "properties" : {
          "Records" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Record"
            }
          }
        }
      },
      "ObjectData" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eTag" : {
            "type" : "string"
          }
        }
      },
      "PatchSampleRequest" : {
        "type" : "object",
        "required" : [ "name", "alias", "mlst", "isolationSource", "collectionDate", "location", "sequencingLab", "institution", "hostHealthState" ],
        "properties" : {
          "name" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "alias" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "taxId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hostTaxId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "mlst" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "isolationSource" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "collectionDate" : {
            "type" : "string",
            "pattern" : "\\S",
            "$ref" : "#/components/schemas/LocalDate"
          },
          "location" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "sequencingLab" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "institution" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "hostHealthState" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        }
      },
      "PresignDownloadRequest" : {
        "type" : "object",
        "required" : [ "projectId", "sampleId", "fileName" ],
        "properties" : {
          "projectId" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "sampleId" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "fileName" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        }
      },
      "PresignResponse" : {
        "type" : "object",
        "properties" : {
          "url" : {
            "type" : "string"
          },
          "objectKey" : {
            "type" : "string"
          },
          "expiresIn" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expiresAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "PresignUploadRequest" : {
        "type" : "object",
        "required" : [ "projectId", "sampleId", "fileName" ],
        "properties" : {
          "projectId" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "sampleId" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "fileName" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        }
      },
      "Project" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "owner" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "createdOn" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "modifiedOn" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "projectMembers" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "$ref" : "#/components/schemas/ProjectMember"
            }
          },
          "samples" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "$ref" : "#/components/schemas/Sample"
            }
          }
        }
      },
      "ProjectMember" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "memberId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "role" : {
            "$ref" : "#/components/schemas/ProjectRole"
          },
          "project" : {
            "$ref" : "#/components/schemas/Project"
          }
        }
      },
      "ProjectResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "owner" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }
      },
      "ProjectRole" : {
        "type" : "string",
        "enum" : [ "OWNER", "ADMIN", "EDITOR", "VIEWER" ]
      },
      "Record" : {
        "type" : "object",
        "properties" : {
          "s3" : {
            "$ref" : "#/components/schemas/S3"
          },
          "eventName" : {
            "type" : "string"
          }
        }
      },
      "S3" : {
        "type" : "object",
        "properties" : {
          "bucket" : {
            "$ref" : "#/components/schemas/Bucket"
          },
          "object" : {
            "$ref" : "#/components/schemas/ObjectData"
          }
        }
      },
      "Sample" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "name" : {
            "type" : "string"
          },
          "alias" : {
            "type" : "string"
          },
          "taxId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hostTaxId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "mlst" : {
            "type" : "string"
          },
          "isolationSource" : {
            "type" : "string"
          },
          "collectionDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "location" : {
            "type" : "string"
          },
          "sequencingLab" : {
            "type" : "string"
          },
          "institution" : {
            "type" : "string"
          },
          "hostHealthState" : {
            "type" : "string"
          },
          "createdOn" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "modifiedOn" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "project" : {
            "$ref" : "#/components/schemas/Project"
          },
          "files" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "$ref" : "#/components/schemas/File"
            }
          }
        }
      },
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
      "UpdateProjectRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "UploadStatus" : {
        "type" : "string",
        "enum" : [ "PENDING", "UPLOADED" ]
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "roles" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "openIdConnectUrl" : "http://localhost:8080/realms/metatrack/.well-known/openid-configuration",
        "description" : "Authentication"
      }
    }
  },
  "paths" : {
    "/api/files/presign-download" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PresignDownloadRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PresignResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Presign Download",
        "tags" : [ "File Controller" ]
      }
    },
    "/api/files/presign-upload" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PresignUploadRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PresignResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Presign Upload",
        "tags" : [ "File Controller" ]
      }
    },
    "/api/projects" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectResponse"
                  }
                }
              }
            }
          }
        },
        "summary" : "Get All Projects",
        "tags" : [ "Project Controller" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create New Project",
        "tags" : [ "Project Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/projects/{projectId}" : {
      "patch" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Update Project",
        "tags" : [ "Project Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Delete Project",
        "tags" : [ "Project Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/projects/{projectId}/member/{memberId}" : {
      "delete" : {
        "parameters" : [ {
          "name" : "memberId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Remove Member",
        "tags" : [ "Project Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "post" : {
        "parameters" : [ {
          "name" : "memberId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Add Member",
        "tags" : [ "Project Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/projects/{projectId}/samples" : {
      "get" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Sample"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get All Samples",
        "tags" : [ "Sample Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "post" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSampleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Sample"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Sample",
        "tags" : [ "Sample Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/projects/{projectId}/samples/name/{sampleName}" : {
      "get" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "sampleName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Sample"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get Sample By Name",
        "tags" : [ "Sample Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/projects/{projectId}/samples/samplesheet" : {
      "post" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Import CSV",
        "tags" : [ "Sample Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/projects/{projectId}/samples/{sampleId}" : {
      "patch" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "sampleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatchSampleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Update Sample",
        "tags" : [ "Sample Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "sampleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Sample"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get Sample",
        "tags" : [ "Sample Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "sampleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Delete Sample",
        "tags" : [ "Sample Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/whoami" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get User Info",
        "tags" : [ "User Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/status" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HealthStatusResponse"
                }
              }
            }
          }
        },
        "summary" : "Status",
        "tags" : [ "Status Controller" ]
      }
    },
    "/webhook/minio" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MinioEvent"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Process Webhook",
        "tags" : [ "Minio Webhook Controller" ]
      }
    }
  },
  "info" : {
    "title" : "server API",
    "version" : "1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:1234",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:1234",
    "description" : "Auto generated value"
  } ]
}